gaussian_splatting: evaluate band-1 SH with the reference view direction; review round 1 for signed SH (#1054, #1063)
Review round 1 of the signed SH storage slice.

- #1063: tile_binning.glsl evaluated the SH basis with dir = camera - splat. The reference
  (Inria 3DGS computeColorFromSH, forward.cu) uses dir = pos - campos. The basis is odd in dir
  for bands 1 and 3, so every trained band-1/band-3 coefficient rendered with the opposite sign.
  The unsigned RGB9E5 clamp had partly hidden this.
- SH words are uint end to end (PackedSphericalHarmonics.encoded, GLSL sh_encoded[12] in all
  four mirrors), so no driver can flush a SNORM10 word that happens to be a float denormal.
- Encoder: a non-finite inverse step (scale below ~1.5e-36) now stores zero words instead of
  converting inf/NaN to int (undefined behaviour).
- The unsupported-encoding counter also counts encoding 0 with a non-zero coefficient count.
- GPU test (RendererSceneTree batch) now asserts the analytic reference delta c * basis_k(dir)
  per band-1 slot and sign, on-axis and off-axis, on the unquantized AND quantized atlas routes,
  and waits with gs_pump_until. The fixture keeps band 1 complete (a green slot-2 coefficient):
  GaussianData derives sh_degree from the highest non-zero band-1 slot and the quantized route
  gates bands by it, so a lone slot-0/1 coefficient rendered DC-only there. Premise asserted.
  Mutation-proven on an RTX 3090 / Vulkan: reversing the view direction fails 14/24
  assertions; clamping negative coefficients to zero fails 7/24.
- Layout guard: sign extension pinned via the signed `int word`, checks scoped to function
  bodies, the sh_dc.w / dc[3] scale lane pinned, encoder shifts and clamp pinned, and the view
  direction convention pinned.
- Release gate manifest: RequiresGPU snapshot 163 -> 164 (recomputed with the repo's own
  functions); RendererSceneTree minimum_test_cases 1 -> 2.
- Stale text: RGB9E5 comments, generated shader reference, the RendererSceneTree batch comment,
  and a test_quantized_packing.h comment citing a PSNR test that does not exist.

The dead second RGB9E5 decoder (gaussian_splat_common_inc.glsl, used only by the never-drawn
gaussian_splat.glsl) stays by maintainer waiver; deletion is a follow-up.

Refs #1054, #1063.

+++ modules/gaussian_splatting/compute/depth_compute.glsl
@@ -27,7 +27,7 @@ struct PackedGaussian {
     uint _pad_rotation_align;  // std430 keeps rotation at offset 32; was `area` (never read)
     vec4 rotation;
     vec4 sh_dc;
-    float sh_encoded[12];
+    uint sh_encoded[12];  // #1054: raw SNORM10 SH words; uint (not float) so no driver can flush them as denormals
     vec3 normal;
     uint sh_metadata;
 };
